+++ app/messages/[threadId]/page.tsx
@@ -2,6 +2,6 @@
 
 import InboxV2 from "@/components/messages/InboxV2";
 
-export default function MessagesPage() {
+export default function MessageThreadPage() {
   return <InboxV2 />;
 }